Refactor stem separation functions and enhance audio device callback handling
- Updated `runStemSeparationJob` function signature for better const correctness.
- Modified `NullStemSeparator` to improve parameter clarity in `separate` method.
- Added regression tests for transport playhead behavior, ensuring play, pause, and resume functionality works as expected.
- Enhanced developer guide with detailed instructions for using clang-tidy, including new parameters for filtering and reporting.
- Updated frontend linting command to simplify ESLint usage.
- Improved `Invoke-ClangTidy.ps1` script to support additional parameters and better error handling.
- Introduced `DeviceCallbackGuard` class to monitor audio device state transitions, ensuring accurate reporting of device status.

+++ backend/src/scratch/ScratchSourcePreparation.cpp
@@ -260,7 +260,7 @@ bool prepareSourceToCache(const SourcePreparationSettings& settings,
     else
     {
         prepared = std::move(source);
-        if (sourceSamples / reader->sampleRate >= kLongUnwarpedClipSeconds)
+        if (static_cast<double>(sourceSamples) / reader->sampleRate >= kLongUnwarpedClipSeconds)
         {
             result.audio =
                 std::make_shared<juce::AudioBuffer<float>>(std::move(prepared));
